Moved Express app with middleware and routes to separate file
