+++ Backend/server.js
@@ -1,17 +1,13 @@
 // Third-party modules
 const fs = require("fs");
 const https = require("https");
-const express = require("express");
-const { v4: uuidv4 } = require("uuid");
 var assert = require("assert");
 
 // Custom application modules
-const app = express();
+const app = require("./app.js");
 const db = require("./Database/dbSetup.js");
 const GameManager = require("./assets/GameManager.js");
 const UserDBManager = require("./assets/UserDBManager.js");
-const Player = require("./assets/Player.js");
-const User = require("./assets/User.js");
 const PlayerAction = require("./assets/PlayerAction.js");
 const { Socket } = require("socket.io");
 
@@ -24,38 +20,6 @@ const certificate = fs.readFileSync("./cert.pem", "utf8");
 
 const credentials = { key: privateKey, cert: certificate };
 const httpsServer = https.createServer(credentials, app);
-
-/**
- * Middleware to validate session token
- *
- * ChatGPT usage: Partial
- */
-app.use(express.json());
-app.use((req, res, next) => {
-  // Exclude the middleware for /login and /create-account
-  if (req.path === "/login" || req.path === "/create-account") {
-    return next();
-  } else {
-    const sessionToken = req.body.sessionToken;
-
-    if (sessionToken) {
-      userDBManager.getUserBySessionToken(sessionToken).then((user) => {
-        if (user) {
-          req.user = user;
-          next();
-        } else {
-          res.status(404).send({
-            message: "Unable to find the user for this account.",
-          });
-        }
-      });
-    } else {
-      res.status(404).send({
-        message: "Unable to find the user for this account.",
-      });
-    }
-  }
-});
 
 /* Starts the server and database */
 const server = httpsServer.listen(8081, "0.0.0.0", async () => {
@@ -74,254 +38,6 @@ const server = httpsServer.listen(8081, "0.0.0.0", async () => {
     gameManager.testing();
     console.log("test room added!");
   }
-});
-
-/**
- * Creates a new user account.
- *
- * This endpoint should be called after the client successfully logs in
- * to their Google Account, in which case the server will add the user to
- * the database.
- *
- * ChatGPT usage: No
- */
-app.post("/create-account", (req, res) => {
-  const token = req.body.token;
-  const username = req.body.username;
-
-  if (!token || !username) {
-    res.status(400).send({ message: "Invalid parameters were passed in." });
-    return;
-  }
-
-  userDBManager.createNewUser(token, username).then(
-    (user) => {
-      console.log(user);
-
-      if (user) {
-        const userString = JSON.stringify({
-          token: user.token,
-          username: user.username,
-          rank: user.rank,
-        });
-
-        res.status(201).send(userString);
-      } else {
-        res
-          .status(500)
-          .send({ message: "There was an error creating the account." });
-      }
-    },
-    (err) => {
-      console.log("[ERROR]: " + err);
-
-      res
-        .status(400)
-        .send({ message: "An account already exists for this user." });
-    }
-  );
-});
-
-/**
- * Logs the user in, generating a new session token for use in future API calls.
- *
- * ChatGPT usage: Partial
- */
-app.post("/login", (req, res) => {
-  const token = req.body.token;
-
-  // Generate and set the user's session token
-  const sessionToken = uuidv4();
-
-  userDBManager.setUserSessionToken(token, sessionToken).then(
-    (user) => {
-      if (user) {
-        res.status(200).send({
-          token: user.token,
-          username: user.username,
-          rank: user.rank,
-          sessionToken: user.sessionToken,
-        });
-      } else {
-        res.status(400).send({
-          message: "Unable to find the user for this account.",
-        });
-      }
-    },
-    (err) => {
-      console.log("[ERROR]: " + err);
-      res.status(500).send({ message: "An unknown error occurred" });
-    }
-  );
-});
-
-/**
- * Logs the user out, destroying their session token.
- *
- * ChatGPT usage: No
- */
-app.post("/logout", (req, res) => {
-  userDBManager.getUserBySessionToken(req.sessionToken).then((loggedInUser) => {
-    userDBManager.setUserSessionToken(loggedInUser.token, null).then(
-      (loggedOutUser) => {
-        assert(loggedOutUser.sessionToken === null);
-        res.status(200).send();
-      },
-      (err) => {
-        console.log("[ERROR]: " + err);
-        res.status(500).send({ message: "An unknown error occurred" });
-      }
-    );
-  });  
-});
-
-/**
- * Puts the user in a random active game room. The user will only
- * be put in a game room that is marked as public.
- *
- * ChatGPT usage: No
- */
-app.post("/join-random-room", (req, res) => {
-  const user = req.user;
-
-  // Fetch all the public game rooms with space remaining for new players
-  const availableRooms = gameManager.getAvailableRooms();
-
-  try {
-    if (availableRooms.length == 0) {
-      res.status(404).send({
-        message: "No game rooms available at the moment. Please try again later.",
-      });
-      return;
-    }
-  
-    // Prioritize rooms that have been waiting for a long time
-    availableRooms.sort(
-      (roomA, roomB) => roomA.getRoomCreationTime() - roomB.getRoomCreationTime()
-    );
-  
-    // Prioritize rooms with the rank of players that are closest to the user's rank
-    const roomPriorities = [];
-  
-    for (let i = 0; i < availableRooms.length; i++) {
-      const room = availableRooms[i];
-  
-      let players = room.getPlayers();
-  
-      // Calculate the average rank
-      let totalRank = players.reduce((sum, player) => sum + player.rank, 0);
-      let averageRank = totalRank / players.length;
-  
-      // This priority is a weighted value that considers how long the room has been
-      // waiting and how similar the ranks of other players in that room are compared
-      // to the user.
-      //
-      // Rooms that have been waiting for a long time will have a lower priority value.
-      // Rooms with average player ranks that are similar to the user will also have
-      // a lower priority value.
-      //
-      // The lower the priority value, the more suitable the room is for the user.
-      const priority = i + Math.abs(user.rank - averageRank);
-  
-      roomPriorities.push({ roomCode: room.roomCode, priority: priority });
-    }
-  
-    roomPriorities.sort((roomA, roomB) => roomA["priority"] - roomB["priority"]);
-  
-    const player = new Player(user);
-  
-    for (var roomPriority of roomPriorities) {
-      let roomCode = roomPriority.roomCode;
-      let room = gameManager.fetchRoom(roomCode);
-
-      let joinSuccess = room.addPlayer(player);
-  
-      if (joinSuccess) {
-        res.status(200).send({
-          roomId: room.roomId,
-          roomCode: room.roomCode,
-        });
-  
-        break;
-      }
-    }
-  } catch (err) {
-    console.log(err);
-    res.status(500).send({message: err});
-  }
-});
-
-/**
- * Allows a user to join an active game room via code.
- *
- * ChatGPT usage: Partial
- */
-app.post("/join-room-by-code", (req, res) => {
-  const user = req.user;
-  const roomCode = req.body.roomCode;
-
-  const room = gameManager.fetchRoom(roomCode);
-
-  try {
-    if (room) {
-      const userBanned = room.isUserBanned(user.username);
-  
-      if (userBanned) {
-        res.status(403).send({ message: "You are banned from this game room." });
-      } else {
-        const player = new Player(user);
-        const joinSuccess = room.addPlayer(player);
-  
-        if (joinSuccess) {
-          res.status(200).send({
-            roomId: room.roomId,
-            roomCode: room.roomCode,
-          });
-        } else {
-          res.status(409).send({
-            message: "The game room is currently full. Please try again later.",
-          });
-        }
-      }
-    } else {
-      res.status(404).send({ message: "The game room could not be found." });
-    }
-  } catch (err) {
-    console.log(err);
-    res.status(500).send({message: err});
-  }
- 
-});
-
-/**
- * Creates a new Game Room for the user, who will be the game master.
- *
- * ChatGPT usage: No
- */
-app.post("/create-room", (req, res) => {
-  const sessionToken = req.body.sessionToken;
-
-  userDBManager.getUserBySessionToken(sessionToken).then(
-    (dbUser) => {
-      const user = new User(
-        dbUser.token,
-        dbUser.username,
-        dbUser.rank,
-        sessionToken
-      );
-      const gameMaster = new Player(user);
-
-      const room = gameManager.createGameRoom(gameMaster);
-
-      res.status(200).send({ roomId: room.roomId });
-    },
-    (err) => {
-      console.log(err);
-      res
-        .status(500)
-        .send({ message: err });
-    }
-  );
 });
 
 // Delay between start of game and question
@@ -412,33 +128,33 @@ io.on("connection", (socket) => {
       // in our welcomeNewPlayer event payload
       const playersJson = [];
       let newPlayerRank;
-  
+
       for (let player of players) {
         playersJson.push({
           username: player.user.username,
           rank: player.user.rank,
           isReady: player.isReady,
         });
-  
+
         if (player.user.username === message.username) {
           // Keep the new player's rank to send in the payload of playerJoined
           newPlayerRank = player.user.rank;
-  
+
           // Since joinRoom is the first event that is emitted by a client
           // after they connect to the socket, we need to store their socket id
           console.log(
             `User ${username} with socket.id=${socket.id} joined room ${room.roomCode}`
           );
-  
+
           player.setSocketId(socket.id);
         }
       }
-  
+
       const roomSettings = room.getSettings();
-  
+
       // Player Joins Room
       socket.join(room.roomId);
-  
+
       // Send Room Data to Player
       socket.emit("welcomeNewPlayer", {
         roomPlayers: playersJson,
@@ -446,7 +162,7 @@ io.on("connection", (socket) => {
         possibleCategories: gameManager.possibleCategories,
         roomCode: room.roomCode,
       });
-  
+
       // Notify players in the room that a new player has joined
       socket.to(message.roomId).emit("playerJoined", {
         newPlayerUsername: username,
@@ -454,9 +170,8 @@ io.on("connection", (socket) => {
       });
     } catch (err) {
       console.log(err);
-      socket.emit("error", {message: message});
+      socket.emit("error", { message: message });
     }
-   
   });
 
   /**
@@ -477,11 +192,11 @@ io.on("connection", (socket) => {
         for (let player of room.getPlayers()) {
           const playerUsername = player.user.username;
           room.removePlayer(playerUsername);
-  
+
           if (player === undefined) {
             continue;
           }
-  
+
           // Be sure to also remove them from this socket room
           let socketId = player.getSocketId();
           if (socketId != undefined) {
@@ -492,26 +207,26 @@ io.on("connection", (socket) => {
             }
           }
         }
-  
+
         // The room should now be empty. Remove the room so that no one
         // can join it.
         assert(room.getPlayers().length === 0);
         const success = gameManager.removeRoomById(roomId);
-  
+
         if (!success) {
           console.log("Could not remove room with id " + room.roomId);
         }
-  
+
         console.log("Room was removed successfully");
       } else {
         const player = room.getPlayer(username);
         room.removePlayer(username);
-  
+
         // Be sure to also remove them from this socket room
         if (player === undefined) {
           return;
         }
-  
+
         let socketId = player.getSocketId();
         if (socketId != undefined) {
           let playerSocket = io.sockets.sockets.get(socketId);
@@ -520,13 +235,13 @@ io.on("connection", (socket) => {
             playerSocket.emit("roomClosed");
           }
         }
-  
+
         // Notify other players still in the room that a player
         // has left
         socket
           .to(roomId)
           .emit("playerLeft", { playerUsername: username, reason: "left" });
-  
+
         // Notify the player who left that their request has been fulfilled.
         socket.emit("removedFromRoom", {
           reason: "left",
@@ -534,9 +249,8 @@ io.on("connection", (socket) => {
       }
     } catch (err) {
       console.log(err);
-      socket.emit("error", {message: message});
+      socket.emit("error", { message: message });
     }
-    
   });
 
   /**
@@ -557,13 +271,13 @@ io.on("connection", (socket) => {
         socket.emit("error", {
           message: "You must be the game room owner to ban another user.",
         });
-  
+
         return;
       }
-  
+
       room.removePlayer(bannedUsername);
       room.banPlayer(bannedUsername);
-  
+
       // Notify other players that a player has been banned from the room
       socket
         .to(roomId)
@@ -571,12 +285,12 @@ io.on("connection", (socket) => {
           "playerLeft",
           express.json({ playerUsername: bannedUsername, reason: "banned" })
         );
-  
+
       // Notify the banned player that they have been banned
       const bannedPlayer = room.getPlayer(bannedUsername);
-  
+
       const bannedPlayerSocketId = bannedPlayer.getSocketId();
-  
+
       if (bannedPlayerSocketId != undefined) {
         let bannedPlayerSocket = io.sockets.sockets.get(bannedPlayerSocketId);
         bannedPlayerSocket.emit("removedFromRoom", {
@@ -585,9 +299,8 @@ io.on("connection", (socket) => {
       }
     } catch (err) {
       console.log(err);
-      socket.emit("error", {message: message});
+      socket.emit("error", { message: message });
     }
-    
   });
 
   /**
@@ -767,9 +480,9 @@ io.on("connection", (socket) => {
     try {
       const room = gameManager.fetchRoomById(roomId);
       const roomCode = room.roomCode;
-  
+
       socket.to(roomId).emit("answerReceived", { playerUsername });
-  
+
       const newAnswer = new PlayerAction(
         message.username,
         message.timeDelay,
@@ -781,16 +494,16 @@ io.on("connection", (socket) => {
         roomCode,
         newAnswer
       );
-  
+
       if (allAnswersReceived) {
         // Get points per round
         const results = gameManager.calculateScore(roomCode);
-  
+
         if (results.returnCode == 0) {
           //Calculate new totals
           const scoreGain = results.scores;
           let totalScores = gameManager.addToPlayerScore(roomCode, scoreGain);
-  
+
           // Format Points per round response and send
           let scores = [];
           totalScores.forEach((score) => {
@@ -801,12 +514,12 @@ io.on("connection", (socket) => {
               updatedTotalPoints: score.finalScore,
             });
           });
-  
+
           const scoresData = { scores };
-  
+
           socket.to(roomId).emit("showScoreboard", scoresData);
           socket.emit("showScoreboard", scoresData);
-  
+
           // If no remaiing questiosns, end game, else send next questions
           if (gameManager.fetchQuestionsQuantity(roomCode) != 0) {
             setTimeout(() => {
@@ -817,60 +530,60 @@ io.on("connection", (socket) => {
               socket.to(roomId).emit("endGame", { scores: totalScores });
               socket.emit("endGame", { scores: totalScores });
             }, SHOW_SCOREBOARD_MILLISECONDS);
-  
+
             // Update ranks in user profile of all players
             let roomPlayers = room.getPlayers();
             let numPlayers = roomPlayers.length;
             let rankValues = [];
-  
+
             // Sort the array of room players from highest to lowest points
             roomPlayers.sort((a, b) => b.points - a.points);
-  
+
             switch (numPlayers) {
               case 2:
                 rankValues = [1, -1];
                 break;
-  
+
               case 3:
                 rankValues = [2, 0, -2];
                 break;
-  
+
               case 4:
                 rankValues = [3, 1, -1, -3];
                 break;
-  
+
               case 5:
                 rankValues = [3, 2, 0, -2, -3];
                 break;
-  
+
               case 6:
                 rankValues = [3, 2, 1, -1, -2, -3];
                 break;
-  
+
               default:
                 for (let i = 0; i < numPlayers; i++) {
                   rankValues.push(0);
                 }
-  
+
                 break;
             }
-  
+
             for (let i = 0; i < numPlayers; i++) {
               let player = roomPlayers[i];
               let value = rankValues[i];
-  
+
               userDBManager.updateUserRank(player.user.username, value);
             }
-  
+
             // Now remove all players from room and delete the room.
             for (let player of room.getPlayers()) {
               const playerUsername = player.user.username;
               room.removePlayer(playerUsername);
-  
+
               if (player === undefined) {
                 continue;
               }
-  
+
               let socketId = player.getSocketId();
               if (socketId != undefined) {
                 let playerSocket = io.sockets.sockets.get(socketId);
@@ -879,10 +592,10 @@ io.on("connection", (socket) => {
                 }
               }
             }
-  
+
             assert(room.getPlayers().length === 0);
             const success = gameManager.removeRoomById(roomId);
-  
+
             if (!success) {
               console.log("Could not remove room with id " + room.roomId);
             }
@@ -891,9 +604,8 @@ io.on("connection", (socket) => {
       }
     } catch (err) {
       console.log(err);
-      socket.emit("error", {message: message});
-    } 
-   
+      socket.emit("error", { message: message });
+    }
   });
 
   /**
